requirements_test.txt
pytest>=7.0.0
pytest-homeassistant-custom-component
pytest-asyncio
requests
python-dotenv
